refactor: extract note actions from App
