+++ src/app/App.tsx
@@ -220,6 +220,8 @@ import { createAppDomainActions } from "./app-domain-actions";
 
 import { useAppUpdateActions } from "./use-app-update-actions";
 
+import { createNoteActions } from "./note-actions";
+
 export function App() {
   const replayOnboarding = shouldReplayOnboarding();
   const activeHermesProfileName = useActiveHermesProfileName();
@@ -2199,82 +2201,19 @@ export function App() {
     setFolderReturnTarget(undefined);
   }
 
-  async function handleDeleteNote(noteId: string) {
-    if (state.recordingStatus) {
-      setError("Stop the current recording before deleting a note.");
-      return;
-    }
-    try {
-      await deleteNote(noteId);
-      pruneDeletedNoteTabs(new Set([noteId]));
-      const response = await listNotes();
-      dispatch({ type: "notesLoaded", notes: response.items });
-      const nextNoteId = response.items[0]?.id;
-      if (nextNoteId) {
-        const note = await getNote(nextNoteId);
-        dispatch({ type: "noteLoaded", note });
-      } else {
-        handleEmptyNotesAfterDelete();
-      }
-    } catch (err) {
-      setError(messageFromError(err));
-    }
-  }
-
-  async function handleDeleteNotes(noteIds: string[]) {
-    if (state.recordingStatus) {
-      setError("Stop the current recording before deleting meetings.");
-      return;
-    }
-    try {
-      await deleteNotes(noteIds);
-      pruneDeletedNoteTabs(new Set(noteIds));
-      const response = await listNotes();
-      dispatch({ type: "notesLoaded", notes: response.items });
-      const nextNoteId = response.items[0]?.id;
-      if (nextNoteId) {
-        const note = await getNote(nextNoteId);
-        dispatch({ type: "noteLoaded", note });
-      } else {
-        handleEmptyNotesAfterDelete();
-      }
-    } catch (err) {
-      setError(messageFromError(err));
-    }
-  }
-
-  async function handleSelectNoteFromFolder(noteId: string, folderId: string) {
-    try {
-      const note = await getNote(noteId);
-      dispatch({ type: "noteLoaded", note });
-      setOriginFolderId(folderId);
-      setOriginAllNotes(false);
-      setFolderReturnTarget(undefined);
-      setActiveView("meetings");
-    } catch (err) {
-      setError(messageFromError(err));
-    }
-  }
-
-  async function handleUpdateNote(patch: Partial<Pick<NoteDto, "title" | "editedContent">>) {
-    if (!selectedNote) return;
-    const optimistic = {
-      ...selectedNote,
-      ...patch,
-      updatedAt: new Date().toISOString(),
-    };
-    dispatch({ type: "noteUpdated", note: optimistic });
-    try {
-      const note = await updateNote({
-        noteId: selectedNote.id,
-        title: patch.title,
-        editedContent: patch.editedContent,
-      });
-      dispatch({ type: "noteUpdated", note });
-    } catch (err) {
-      setError(messageFromError(err));
-    }
-  }
+  const { handleDeleteNote, handleDeleteNotes, handleSelectNoteFromFolder, handleUpdateNote } =
+    createNoteActions({
+      dispatch,
+      handleEmptyNotesAfterDelete,
+      pruneDeletedNoteTabs,
+      selectedNote,
+      setActiveView,
+      setError,
+      setFolderReturnTarget,
+      setOriginAllNotes,
+      setOriginFolderId,
+      state,
+    });
 
   const {
     handleStartRecording,
