+++ src/app/note-actions-types.ts
@@ -1,0 +1,19 @@
+import { type SidebarView } from "../components/sidebar/Sidebar";
+import type { NoteDto } from "../lib/tauri";
+import type { NotesAction, NotesState } from "./state/app-state";
+import type * as React from "react";
+
+export type CreateNoteActionsDependencies = {
+  dispatch: React.Dispatch<NotesAction>;
+  handleEmptyNotesAfterDelete: () => void;
+  pruneDeletedNoteTabs: (removedIds: Set<string>) => void;
+  selectedNote: NoteDto | undefined;
+  setActiveView: React.Dispatch<React.SetStateAction<SidebarView>>;
+  setError: React.Dispatch<React.SetStateAction<string | null>>;
+  setFolderReturnTarget: React.Dispatch<
+    React.SetStateAction<{ noteId: string; label: string } | undefined>
+  >;
+  setOriginAllNotes: React.Dispatch<React.SetStateAction<boolean>>;
+  setOriginFolderId: React.Dispatch<React.SetStateAction<string | undefined>>;
+  state: NotesState;
+};
