+++ src/app/note-actions.ts
@@ -1,0 +1,103 @@
+import { deleteNote, deleteNotes, getNote, listNotes, updateNote } from "../lib/tauri";
+import { messageFromError } from "../lib/errors";
+import type { NoteDto } from "../lib/tauri";
+import type { CreateNoteActionsDependencies } from "./note-actions-types";
+
+export function createNoteActions(dependencies: CreateNoteActionsDependencies) {
+  const {
+    dispatch,
+    handleEmptyNotesAfterDelete,
+    pruneDeletedNoteTabs,
+    selectedNote,
+    setActiveView,
+    setError,
+    setFolderReturnTarget,
+    setOriginAllNotes,
+    setOriginFolderId,
+    state,
+  } = dependencies;
+
+  async function handleDeleteNote(noteId: string) {
+    if (state.recordingStatus) {
+      setError("Stop the current recording before deleting a note.");
+      return;
+    }
+    try {
+      await deleteNote(noteId);
+      pruneDeletedNoteTabs(new Set([noteId]));
+      const response = await listNotes();
+      dispatch({ type: "notesLoaded", notes: response.items });
+      const nextNoteId = response.items[0]?.id;
+      if (nextNoteId) {
+        const note = await getNote(nextNoteId);
+        dispatch({ type: "noteLoaded", note });
+      } else {
+        handleEmptyNotesAfterDelete();
+      }
+    } catch (err) {
+      setError(messageFromError(err));
+    }
+  }
+
+  async function handleDeleteNotes(noteIds: string[]) {
+    if (state.recordingStatus) {
+      setError("Stop the current recording before deleting meetings.");
+      return;
+    }
+    try {
+      await deleteNotes(noteIds);
+      pruneDeletedNoteTabs(new Set(noteIds));
+      const response = await listNotes();
+      dispatch({ type: "notesLoaded", notes: response.items });
+      const nextNoteId = response.items[0]?.id;
+      if (nextNoteId) {
+        const note = await getNote(nextNoteId);
+        dispatch({ type: "noteLoaded", note });
+      } else {
+        handleEmptyNotesAfterDelete();
+      }
+    } catch (err) {
+      setError(messageFromError(err));
+    }
+  }
+
+  async function handleSelectNoteFromFolder(noteId: string, folderId: string) {
+    try {
+      const note = await getNote(noteId);
+      dispatch({ type: "noteLoaded", note });
+      setOriginFolderId(folderId);
+      setOriginAllNotes(false);
+      setFolderReturnTarget(undefined);
+      setActiveView("meetings");
+    } catch (err) {
+      setError(messageFromError(err));
+    }
+  }
+
+  async function handleUpdateNote(patch: Partial<Pick<NoteDto, "title" | "editedContent">>) {
+    if (!selectedNote) return;
+    const optimistic = {
+      ...selectedNote,
+      ...patch,
+      updatedAt: new Date().toISOString(),
+    };
+    dispatch({ type: "noteUpdated", note: optimistic });
+    try {
+      const note = await updateNote({
+        noteId: selectedNote.id,
+        title: patch.title,
+        editedContent: patch.editedContent,
+      });
+      dispatch({ type: "noteUpdated", note });
+    } catch (err) {
+      setError(messageFromError(err));
+    }
+  }
+
+  return {
+    handleDeleteNote,
+    handleDeleteNotes,
+    handleSelectNoteFromFolder,
+    handleUpdateNote,
+  };
+}
